Cleaning Up and Adding A2P4 Functionality
Cleaned up the relic of the first assignment submission. Made a new project withing the solution linking to the main project to be able to have distinct entry points for the demo. Change the player resolve dice. Now with an set ordered in the way the player is resolving his dice roll a map is produce that associate the Die::Face with the particular amount of that symbol was rolled. Need to include the option to reroll up to three times. Added a PlayerStateMachine which will be use to keep track of the state of a player through out the game.